Crate termion

Termion is a pure Rust, bindless library for low-level handling, manipulating and reading information about terminals. This provides a full-featured alternative to Termbox.

Termion aims to be simple and yet expressive. It is bindless, meaning that it is not a front-end to some other library (e.g., ncurses or termbox), but a standalone library directly talking to the TTY.

Supports Redox, Mac OS X, and Linux (or, in general, ANSI terminals).

For more information refer to the README.

Modules

clear

Clearing the screen.

color

Color managemement.

cursor

Cursor movement.

event

Mouse and key events.

input

User input.

raw

Managing raw mode.

screen

Managing switching between main and alternate screen buffers.

scroll

Scrolling.

style

Text styling management.

Structs

AsyncReader

An asynchronous reader.

Functions

async_stdin

Construct an asynchronous handle to the TTY standard input.

get_tty

Get the TTY device.

is_tty

Is this stream a TTY?

terminal_size

Get the size (columns, rows) of the terminal.

terminal_size_pixels

Get the size of the terminal in pixels.