Enum termion::color::Palette

pub enum Palette {
    Black,
    Red,
    Green,
    Yellow,
    Blue,
    Magenta,
    Cyan,
    White,
    LightBlack,
    LightRed,
    LightGreen,
    LightYellow,
    LightBlue,
    LightMagenta,
    LightCyan,
    LightWhite,
    Rgb(u8, u8, u8),
    Grayscale(u8),
}

A color palette.

This should generally only be used when the color is runtime determined. Otherwise, use the color types, which resolves the value at compile time.

Variants
Black

Black.

Red

Red.

Green

Green.

Yellow

Yellow.

Blue

Blue.

Magenta

Megenta.

Cyan

Cyan.

White

White.

LightBlack

High-intensity black.

LightRed

High-intensity red.

LightGreen

High-intensity green.

LightYellow

High-intensity yellow.

LightBlue

High-intensity blue.

LightMagenta

High-intensity magenta.

LightCyan

High-intensity cyan.

LightWhite

High-intensity white.

Rgb(u8, u8, u8)

216-color (r, g, b ≤ 5) RGB.

Grayscale(u8)

Grayscale (max value: 24).