Trait termion::raw::IntoRawMode

pub trait IntoRawMode: Write + Sized {
    fn into_raw_mode(self) -> Result<RawTerminal<Self>>;
}

Types which can be converted into "raw mode".

Why is this type defined on writers and not readers?

TTYs has their state controlled by the writer, not the reader. You use the writer to clear the screen, move the cursor and so on, so naturally you use the writer to change the mode as well.

Required methods

fn into_raw_mode(self) -> Result<RawTerminal<Self>>

Switch to raw mode.

Raw mode means that stdin won't be printed (it will instead have to be written manually by the program). Furthermore, the input isn't canonicalised or buffered (that is, you can read from stdin one byte of a time). The output is neither modified in any way.

Loading content...