Trait termion::TermWrite

pub trait TermWrite {
    fn csi(&mut self, b: &[u8]) -> Result<usize>;
    fn osc(&mut self, b: &[u8]) -> Result<usize>;
    fn dsc(&mut self, b: &[u8]) -> Result<usize>;

    fn clear(&mut self) -> Result<usize> { ... }
    fn clear_after(&mut self) -> Result<usize> { ... }
    fn clear_before(&mut self) -> Result<usize> { ... }
    fn clear_line(&mut self) -> Result<usize> { ... }
    fn clear_until_newline(&mut self) -> Result<usize> { ... }
    fn show_cursor(&mut self) -> Result<usize> { ... }
    fn hide_cursor(&mut self) -> Result<usize> { ... }
    fn move_cursor_left(&mut self, num: u32) -> Result<usize> { ... }
    fn move_cursor_right(&mut self, num: u32) -> Result<usize> { ... }
    fn reset(&mut self) -> Result<usize> { ... }
    fn restore(&mut self) -> Result<usize> { ... }
    fn goto(&mut self, x: u16, y: u16) -> Result<usize> { ... }
    fn rendition(&mut self, r: u8) -> Result<usize> { ... }
    fn color<C: Color>(&mut self, color: C) -> Result<usize> { ... }
    fn bg_color<C: Color>(&mut self, color: C) -> Result<usize> { ... }
    fn style(&mut self, mode: Style) -> Result<usize> { ... }
}

Extension to the Write trait.

This extension to the Write trait is capable of producing the correct ANSI escape sequences for given commands, effectively controlling the terminal.

Required Methods

fn csi(&mut self, b: &[u8]) -> Result<usize>

Print the CSI (control sequence introducer) followed by a byte string.

fn osc(&mut self, b: &[u8]) -> Result<usize>

Print OSC (operating system command) followed by a byte string.

fn dsc(&mut self, b: &[u8]) -> Result<usize>

Print DSC (device control string) followed by a byte string.

Provided Methods

fn clear(&mut self) -> Result<usize>

Clear the entire screen.

fn clear_after(&mut self) -> Result<usize>

Clear everything after the cursor.

fn clear_before(&mut self) -> Result<usize>

Clear everything before the cursor.

fn clear_line(&mut self) -> Result<usize>

Clear the current line.

fn clear_until_newline(&mut self) -> Result<usize>

Clear from the cursor until newline.

fn show_cursor(&mut self) -> Result<usize>

Show the cursor.

fn hide_cursor(&mut self) -> Result<usize>

Hide the cursor.

fn move_cursor_left(&mut self, num: u32) -> Result<usize>

Move the cursor num spaces to the left.

fn move_cursor_right(&mut self, num: u32) -> Result<usize>

Move the cursor num spaces to the right.

fn reset(&mut self) -> Result<usize>

Reset the rendition mode.

This will reset both the current style and color.

fn restore(&mut self) -> Result<usize>

Restore the defaults.

This will reset color, position, cursor state, and so on. It is recommended that you use this before you exit your program, to avoid messing up the user's terminal.

fn goto(&mut self, x: u16, y: u16) -> Result<usize>

Go to a given position.

The position is 0-based.

fn rendition(&mut self, r: u8) -> Result<usize>

Set graphic rendition.

fn color<C: Color>(&mut self, color: C) -> Result<usize>

Set foreground color.

fn bg_color<C: Color>(&mut self, color: C) -> Result<usize>

Set background color.

fn style(&mut self, mode: Style) -> Result<usize>

Set rendition mode (SGR).