Crate termion

Termion is a pure Rust, bindless library for low-level handling, manipulating and reading information about terminals. This provides a full-featured alternative to Termbox.

Termion aims to be simple and yet expressive. It is bindless, meaning that it is not a front-end to some other library (e.g., ncurses or termbox), but a standalone library directly talking to the TTY.

Supports Redox, Mac OS X, and Linux (or, in general, ANSI terminals).

For more information refer to the README.

Reexports

pub use color::Palette as Color;

Modules

color
ANSI colors.

Structs

AsyncReader
An asynchronous reader.

Keys
An iterator over input keys.

RawTerminal
A terminal restorer, which keeps the previous state of the terminal, and restores it, when dropped.

Enums

Key
A key.

Style
A SGR parameter (rendition mode).

Traits

IntoRawMode
Types which can be converted into "raw mode".

TermRead
Extension to Read trait.

TermWrite
Extension to the Write trait.

Functions

async_stdin
Construct an asynchronous handle to the standard input.

terminal_size
Get the size of the terminal.