Crate termion

termion is a pure Rust library for reading, manipulating, and handling terminals.

This crate is not stable, yet. However, if you do want stability, you should specify the revision (commit hash) in your Cargo.toml, this way builds are complete reproducible, and won't break.

Structs

AsyncReader
An asynchronous reader.

Keys
An iterator over input keys.

RawTerminal
A terminal restorer, which keeps the previous state of the terminal, and restores it, when dropped.

Enums

Color
A terminal color.

Key
A key.

Style
A SGR parameter (rendition mode).

Traits

IntoRawMode
Types which can be converted into "raw mode".

TermRead
Extension to Read trait.

TermWrite
Extension to the Write trait.

Functions

async_stdin
Construct an asynchronous handle to the standard input.

terminal_size
Get the size of the terminal.