Trait termion::color::DetectColors

pub trait DetectColors {
    fn available_colors(&mut self) -> Result<u16>;
}

Types that allow detection of the colors they support.

Required methods

fn available_colors(&mut self) -> Result<u16>

How many ANSI colors are supported (from 8 to 256)?

Beware: the information given isn't authoritative, it's infered through escape codes or the value of TERM, more colors may be available.

Loading content...