Module termion::cursor

Cursor movement.

Structs

BlinkingBar
Change the cursor style to blinking bar

BlinkingBlock
Change the cursor style to blinking block

BlinkingUnderline
Change the cursor style to blinking underline

Down
Move cursor down.

Goto
Goto some position ((1,1)-based).

Hide
Hide the cursor.

HideCursor
Hide the cursor for the lifetime of this struct. It will hide the cursor on creation with from() and show it back on drop().

Left
Move cursor left.

Restore
Restore the cursor.

Right
Move cursor right.

Save
Save the cursor.

Show
Show the cursor.

SteadyBar
Change the cursor style to steady bar

SteadyBlock
Change the cursor style to steady block

SteadyUnderline
Change the cursor style to steady underline

Up
Move cursor up.

Traits

DetectCursorPos
Types that allow detection of the cursor position.