Trait termion::input::TermReadEventsAndRaw

pub trait TermReadEventsAndRaw {
    fn events_and_raw(self) -> EventsAndRaw<Self>
    where
        Self: Sized;
}

Extension to TermRead trait. A separate trait in order to maintain backwards compatibility.

Required methods

fn events_and_raw(self) -> EventsAndRaw<Self> where
Self: Sized,

An iterator over input events and the bytes that define them.