Module termion::raw

Managing raw mode.

Raw mode is a particular state a TTY can have. It signifies that:

No line buffering (the input is given byte-by-byte).

The input is not written out, instead it has to be done manually by the programmer.

The output is not canonicalized (for example, \n means "go one line down", not "line break").

It is essential to design terminal programs.

Example

use termion::raw::IntoRawMode;
use std::io::{Write, stdout};

fn main() {
    let mut stdout = stdout().into_raw_mode().unwrap();

    write!(stdout, "Hey there.").unwrap();
}

Structs

RawTerminal
A terminal restorer, which keeps the previous state of the terminal, and restores it, when dropped.

Constants

CONTROL_SEQUENCE_TIMEOUT
The timeout of an escape code control sequence, in milliseconds.

Traits

IntoRawMode
Types which can be converted into "raw mode".